SCBCRSE rows with the correct EFF_TERM are now selected.
Finding most-recent courses and joining sections to the correct
course rows based on effective term turned out to require using
an 'outer self exclusion join' in the SQL. For the section joins
I put this logic in a new view in the database to ensure that it
would be consistently loaded by section queries from several different
classes.

I added tests to ensure that the correct version of the title and description
are loaded for each section and for the most recent course. This in
turn required the test scripts to ensure that they had access to extra
memory.