Views now use temporary-table algorithm explicitly to improve robustness.
According to the MySQL view documentation http://dev.mysql.com/doc/refman/5.1/en/view-algorithms.html

"A reason to choose TEMPTABLE explicitly is that locks can be released on
underlying tables after the temporary table has been created and before it
is used to finish processing the statement. This might result in quicker
lock release than the MERGE algorithm so that other clients that use the
view are not blocked as long."

I was running into problems if I didn't use the TEMPTABLE algorithm because
section-lookups that occurred while the view was looping through results could
not complete due to locks on the underlying tables.

Unless the List implementation is change to fetch and free all results before
looping, the TEMPTABLE algorithm will likely be required.