For recent courses, we are now fetching the most recent description.
I missed the case where a course effective date is several years ago
since it hadn't needed updating, but a new description was added with
a more recent effective date. In this case there is no reason not to
use the recent description.

In contrast, for sections, we will continue to only use the description
if the effective dates coincide with the term of the section.