Using getDisplayName() as an existence check for apc_course_CourseOffering.
This existence test (getDisplayName) shouldn't really be needed,
but the apc_course_CourseOffering_Lookup_Session::getCourseOffering($id)
method blindly returns course offerings without checking their
existence. While this is a huge performance boost, it also means
that the CourseOffering returned might not be fully usable.

A better fix might be for the apc_course_CourseOfferingLookupSession to do
the test always and cache the result, but APC is broken on my dev machine
at the moment and this is a easier fix to verify the operation of.