Sync: Always running with PDO::ERRMODE_EXCEPTION.
It seems that at least on some versions of MySQL, mis-matched column lengths
and other errors now won't result in silent truncation of data but rather cause
the insert to fail completely, resulting in missing data. If this is going to
happen, it is better to fail and roll-back (with alert emails sent) than continue
with broken data.