#47: Convert id-subcomponent separator from '/' to '.'.
The use of '/' as a separator in subcomponents turned out to be a bad idea
because these slashes need to be escaped as %2F, but these escaped slashes
can interfere with some webserver configuration.

For example Apache redirects require the NE "No Escape" option to preserve
%2F slashes without double-escaping:

    # Redirect all traffic to https
    #
    # We're using mod_rewrite here rather than mod_alias's `Redirect`
    # because mod_alias doesn't allow the "NE" == "noescape" parameter
    # that allows URLs includinging encoded slashes (%2F) to be returned
    # without double-escaping.
    RewriteRule .* https://<%= node["app_domain"] %>$0 [L,R=301,NE]

Similarly, Symfony refuses to allow slashes in parameters passed as tokens
in routes.

To address this problematic choice, use '.' as a separator of
sub-components in identifiers. Periods don't require special escaping in
URLs or database queries, though they do in regular expressions.