Banner: Match catalog for offering by COLL_CODE in addition to TERM_CODE
We were returning too many catalogs for offerings when the offerings' term
codes are used by multiple catalogs.

For example, 202490 is used by sections in both our MCUG catalog as well
as our SA catalog. The term code alone is not enough to determine the
appropriate catalog to use for an Offering/section.