#47: Add a static copy of jquery-ui 1.14.1 for schedule dialogs.
Unfortunately, the normal way of using AssetMapper to include it as a
module was failing because jQuery isn't registered in the global scope
at the point that jquery-ui gets imported. My work-around was to add a line
at the top of jquery-ui.js to import jquery and make the variable available
to the library for modification. This is a bit of a hack, but get both
libraries loaded and available for now.