#47: Convert common uncaught osid exceptions to HTTP exceptions centrally.
osid_NotFoundException are converted to NotFoundHttpException and give a
404 status code to clients.

osid_InvalidArgumentException are converted to BadRequestHttpException and
give a 400 status code to clients.

These central conversions of exceptions uncaught in the controller means
that the controller methods don't have to do as much exception handling
except for cases where they want to ignore certain exceptions and continue.