#47: Use path() instead of url() for local links.
This is a bit of a work-around so that the external link style doesn't get
added to local paths when not hosted under middlebury.edu (such as in dev).
While the midd-frontend's typograpy styles could be updated to also
consider *.lndo.site a non-external link, using relative paths rather than
absolute URLs feels like good hygiene for links anyway to help with
portability.